:art: delete_file_from_ncp moved to utils for flexibility
